Customize for tmf
It shouldn't be difficult in theory to start displaying while the data
is streaming in. The problem is that the page we want to display by
default is at the end of the stream.

An easy way to solve this would be to reverse the order in which the
steps are stored.